Add support for parsing Stmt, restructure parser module
